move committee registration to sui validator tool
