+++ crates/sui-bridge/src/tools/cli.rs
@@ -1,18 +1,11 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use anyhow::anyhow;
 use clap::*;
-use fastcrypto::encoding::{Base64, Encoding};
-use fastcrypto::secp256k1::{Secp256k1KeyPair, Secp256k1PrivateKey};
-use fastcrypto::traits::{KeyPair, ToFromBytes};
-use move_core_types::identifier::Identifier;
 use shared_crypto::intent::Intent;
 use shared_crypto::intent::IntentMessage;
-use std::fs;
 use std::sync::Arc;
 use sui_bridge::client::bridge_authority_aggregator::BridgeAuthorityAggregator;
-use sui_bridge::config::BridgeNodeConfig;
 use sui_bridge::eth_transaction_builder::build_eth_transaction;
 use sui_bridge::sui_client::SuiClient;
 use sui_bridge::sui_transaction_builder::build_sui_transaction;
@@ -23,16 +16,11 @@ use sui_bridge::utils::{
     generate_bridge_authority_key_and_write_to_file, generate_bridge_client_key_and_write_to_file,
     generate_bridge_node_config_and_write_to_file,
 };
-use sui_config::{Config, NodeConfig};
-use sui_json_rpc_types::SuiTransactionBlockEffectsAPI;
-use sui_json_rpc_types::SuiTransactionBlockResponseOptions;
-use sui_sdk::{SuiClient as SuiSdkClient, SuiClientBuilder};
-use sui_types::base_types::SuiAddress;
-use sui_types::bridge::{BridgeChainId, BRIDGE_MODULE_NAME};
+use sui_config::Config;
+use sui_sdk::SuiClient as SuiSdkClient;
+use sui_types::bridge::BridgeChainId;
 use sui_types::crypto::Signature;
-use sui_types::programmable_transaction_builder::ProgrammableTransactionBuilder;
-use sui_types::transaction::{ObjectArg, Transaction, TransactionData};
-use sui_types::BRIDGE_PACKAGE_ID;
+use sui_types::transaction::Transaction;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
@@ -165,85 +153,6 @@ async fn main() -> anyhow::Result<()> {
             };
 
             return Ok(());
-        }
-        BridgeValidatorCommand::CommitteeRegistration {
-            node_config_path,
-            bridge_node_config_path,
-            url,
-        } => {
-            let node_config = NodeConfig::load(node_config_path).expect("Couldn't load NodeConfig");
-            let bridge_config = BridgeNodeConfig::load(bridge_node_config_path)
-                .expect("Couldn't load BridgeNodeConfig");
-
-            // Read bridge keypair
-            let key_str = fs::read_to_string(bridge_config.bridge_authority_key_path_base64_raw)?;
-            let bytes = Base64::decode(key_str.trim()).unwrap();
-            let key = Secp256k1PrivateKey::from_bytes(&bytes).unwrap();
-            let ecdsa_keypair = Secp256k1KeyPair::from(key);
-
-            // Read sui node account key pair
-            let keypair = node_config.account_key_pair.keypair();
-            let address = SuiAddress::from(&keypair.public());
-            println!("Starting bridge committee registration for Sui validator: {address}, with bridge public key: {}", ecdsa_keypair.public);
-
-            let sui_client = SuiClientBuilder::default()
-                .build(&bridge_config.sui_rpc_url)
-                .await?;
-
-            let bridge_client = SuiClient::new(&bridge_config.sui_rpc_url).await?;
-            let bridge = bridge_client
-                .get_mutable_bridge_object_arg()
-                .await
-                .expect("Cannot retrieve bridge object arg.");
-
-            let coins = sui_client
-                .coin_read_api()
-                .get_coins(address, None, None, None)
-                .await?;
-            let gas = coins.data.first().unwrap().object_ref();
-            let ref_gas_price = sui_client.read_api().get_reference_gas_price().await?;
-
-            let mut ptb = ProgrammableTransactionBuilder::default();
-
-            let bridge_arg = ptb.obj(bridge)?;
-            let system_arg = ptb.obj(ObjectArg::SUI_SYSTEM_MUT)?;
-            let pub_key_arg = ptb.pure(ecdsa_keypair.public().as_bytes())?;
-            let url_arg = ptb.pure(url)?;
-            ptb.programmable_move_call(
-                BRIDGE_PACKAGE_ID,
-                BRIDGE_MODULE_NAME.into(),
-                Identifier::new("committee_registration").unwrap(),
-                vec![],
-                vec![bridge_arg, system_arg, pub_key_arg, url_arg],
-            );
-
-            let tx_data = TransactionData::new_programmable(
-                address,
-                vec![gas],
-                ptb.finish(),
-                10000000,
-                ref_gas_price,
-            );
-            let tx = Transaction::from_data_and_signer(tx_data, vec![keypair]);
-            let response = sui_client
-                .quorum_driver_api()
-                .execute_transaction_block(
-                    tx,
-                    SuiTransactionBlockResponseOptions::new().with_effects(),
-                    None,
-                )
-                .await?;
-            if response.status_ok().unwrap() {
-                println!(
-                    "Committee registration successful. txn digest: {}",
-                    response.digest
-                )
-            } else {
-                return Err(anyhow!(
-                    "Error returned from bridge committee registration transaction: {:?}.",
-                    response.effects.as_ref().map(|e| e.status())
-                ));
-            }
         }
     }
