+++ crates/sui-bridge/src/tools/mod.rs
@@ -59,15 +59,6 @@ pub enum BridgeValidatorCommand {
         #[clap(subcommand)]
         cmd: GovernanceClientCommands,
     },
-    CommitteeRegistration {
-        /// Path of BridgeNodeConfig
-        #[clap(long = "node-config-path")]
-        node_config_path: PathBuf,
-        #[clap(long = "bridge-node-config-path")]
-        bridge_node_config_path: PathBuf,
-        #[clap(long = "bridge-authority-url")]
-        url: String,
-    },
 }
 
 #[derive(Parser)]
